[리팩토링] 좋아요 count
LikeEntityRepository 에서 기존에 JPQL을 이용했으나 JPA내에서도 Count 쿼리가 존재함으로 변경하였음

+++ src/main/java/com/example/sns/service/PostService.java
@@ -104,7 +104,7 @@ public class PostService {
         );
     }
 
-    public Integer likeCount(Integer postId) {
+    public long likeCount(Integer postId) {
 
         PostEntity postEntity = getPostEntityOrException(postId);
